+++ devices/MIMXRT1042/drivers/fsl_tempmon.c
@@ -1,6 +1,5 @@
 /*
- * Copyright  2018-2021 NXP
- * All rights reserved.
+ * Copyright  2018-2021, 2024 NXP
  *
  *
  * SPDX-License-Identifier: BSD-3-Clause
@@ -134,7 +133,7 @@ float TEMPMON_GetCurrentTemperature(TEMPMON_Type *base)
     /* Check arguments */
     assert(NULL != base);
 
-    uint32_t nmeas;
+    int32_t nmeas;
     float tmeas;
 
     while (0U == (base->TEMPSENSE0 & TEMPMON_TEMPSENSE0_FINISHED_MASK))
@@ -157,7 +156,7 @@ float TEMPMON_GetCurrentTemperature(TEMPMON_Type *base)
  * param tempVal The alarm temperature with degrees Celsius
  * param alarmMode The alarm mode.
  */
-void TEMPMON_SetTempAlarm(TEMPMON_Type *base, int8_t tempVal, tempmon_alarm_mode alarmMode)
+void TEMPMON_SetTempAlarm(TEMPMON_Type *base, int16_t tempVal, tempmon_alarm_mode alarmMode)
 {
     /* Check arguments */
     assert(NULL != base);
